+++ packages/orq-rc/src/models/operations/generateconversationname.ts
@@ -251,7 +251,7 @@ export const GenerateConversationNameResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kfbcvk33e8azw4k1wm65w2ag"),
+  _id: z.string().default("conv_01kfbeeg16yqa9k4fkt1d8g8yx"),
   kind: GenerateConversationNameKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),
@@ -283,7 +283,7 @@ export const GenerateConversationNameResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   GenerateConversationNameResponseBody
 > = z.object({
-  id: z.string().default("conv_01kfbcvk33e8azw4k1wm65w2ag"),
+  id: z.string().default("conv_01kfbeeg16yqa9k4fkt1d8g8yx"),
   kind: GenerateConversationNameKind$outboundSchema,
   displayName: z.string(),
   createdAt: z.number(),