+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -4569,7 +4569,7 @@ export function agentToolInputRunSchemaFromJSON(
 /** @internal */
 export const Tools$inboundSchema: z.ZodType<Tools, z.ZodTypeDef, unknown> = z
   .object({
-    id: z.string().default("01KFBCVJXJDTX0HTESV7F1K2VJ"),
+    id: z.string().default("01KFBEEFWAD3RY77WNJGR7Y7J3"),
     name: z.string(),
     description: z.string().optional(),
     schema: z.lazy(() => AgentToolInputRunSchema$inboundSchema),
@@ -4588,7 +4588,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01KFBCVJXJDTX0HTESV7F1K2VJ"),
+  id: z.string().default("01KFBEEFWAD3RY77WNJGR7Y7J3"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),