+++ packages/orq-rc/RELEASES.md
@@ -5969,3 +5969,13 @@ Based on:
 - [typescript v4.2.0-rc.43] packages/orq-rc
 ### Releases
 - [NPM v4.2.0-rc.43] https://www.npmjs.com/package/@orq-ai/node/v/4.2.0-rc.43 - packages/orq-rc
+
+## 2026-01-19 15:39:50
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.688.1 (2.796.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.2.0-rc.44] packages/orq-rc
+### Releases
+- [NPM v4.2.0-rc.44] https://www.npmjs.com/package/@orq-ai/node/v/4.2.0-rc.44 - packages/orq-rc